fix(objects)!: model Event_Detection_Enable on the remaining seven intrinsic-reporting types (#216)
Analog Input/Output/Value, Binary Input/Value and Multi-state Input/Value
gained Event_Detection_Enable (ASHRAE 135-2020 Clauses 12.2, 12.3, 12.4,
12.6, 12.8, 12.18, 12.20). Binary Output and Multi-state Output already had
it, so all nine intrinsically-reporting types now model the property.

The property gates both intrinsic-reporting entry points through a new
four-ident arm of impl_intrinsic_reporting!, and a write of FALSE
establishes the Clause 13.2.2.1 initial conditions for the state each type
carries. The analog types reset Event_Time_Stamps and Event_Message_Texts
as well, being the only three that model them; the other four do not model
those properties at all, which is a pre-existing gap tracked as #235.

The default is TRUE, and deliberately differs from Binary Output and
Multi-state Output, which stay FALSE. The standard specifies no default.
The argument is not that the detectors are wholly inert: fault_step() runs
ahead of the event algorithm in both probe() and tick(), keyed on
Reliability alone, and is gated by neither LimitEnable::NONE nor an empty
alarm_values. It is that these seven previously used the *ungated* macro
arm, so that path always ran. TRUE preserves exactly the prior behavior,
where FALSE would have silently suppressed a fault path that works today on
Analog Input/Output/Value. A FALSE default would also have removed all
seven from GetAlarmSummary, GetEnrollmentSummary and GetEventInformation,
since Clause 13.12 makes absence mean included but a present FALSE mean
excluded.

binary/mod.rs is split into input.rs, output.rs and value.rs behind a thin
module root, matching how analog/ and multistate/ are already laid out. The
file had reached 705 LOC against the repo's 700 cap; it was also the only
one of the three modules keeping every object type in a single file.
Behavioral equivalence was checked by a normalized differential of the
extracted match arms, property lists and writability disjuncts, not by
inspection.

BREAKING: the three-ident ungated arm of the exported
impl_intrinsic_reporting! macro is removed. It had no callers left and
exporting an ungated form offers downstream implementors a supported way
to wire detection permanently on.

Fixes #216
