+++ crates/bacnet-objects/src/analog/input.rs
@@ -22,6 +22,9 @@ pub struct AnalogInputObject {
     /// Set to a positive value for delta-based filtering.
     cov_increment: f32,
     event_detector: OutOfRangeDetector,
+    /// Event_Detection_Enable (Clause 12.2). Clause 13.2.2.1: "If the
+    /// Event_Detection_Enable property is FALSE, then this state machine is not evaluated."
+    event_detection_enable: bool,
     /// Reliability: 0 = NO_FAULT_DETECTED.
     reliability: u32,
     /// Optional minimum present value for fault detection.
@@ -48,6 +51,7 @@ impl AnalogInputObject {
             status_flags: StatusFlags::empty(),
             cov_increment: 0.0,
             event_detector: OutOfRangeDetector::default(),
+            event_detection_enable: true,
             reliability: 0,
             min_pres_value: None,
             max_pres_value: None,
@@ -117,6 +121,9 @@ impl BACnetObject for AnalogInputObject {
         if let Some(result) = read_generic_event_properties!(self, property) {
             return result;
         }
+        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
+            return Ok(PropertyValue::Boolean(self.event_detection_enable));
+        }
         match property {
             p if p == PropertyIdentifier::OBJECT_TYPE => {
                 Ok(PropertyValue::Enumerated(ObjectType::ANALOG_INPUT.to_raw()))
@@ -183,6 +190,25 @@ impl BACnetObject for AnalogInputObject {
         {
             return result;
         }
+        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
+            if let PropertyValue::Boolean(v) = value {
+                self.event_detection_enable = v;
+                if !v {
+                    self.event_detector.event_state = bacnet_types::enums::EventState::NORMAL;
+                    self.event_detector.acked_transitions = 0b111;
+                    self.event_detector.pending = None;
+                    self.event_detector.fault_reliability = None;
+                    self.event_time_stamps = [
+                        BACnetTimeStamp::SequenceNumber(0),
+                        BACnetTimeStamp::SequenceNumber(0),
+                        BACnetTimeStamp::SequenceNumber(0),
+                    ];
+                    self.event_message_texts = [String::new(), String::new(), String::new()];
+                }
+                return Ok(());
+            }
+            return Err(common::invalid_data_type_error());
+        }
         if let Some(result) = write_analog_event_properties!(self, property, value) {
             return result;
         }
@@ -201,6 +227,7 @@ impl BACnetObject for AnalogInputObject {
             PropertyIdentifier::PRESENT_VALUE,
             PropertyIdentifier::STATUS_FLAGS,
             PropertyIdentifier::EVENT_STATE,
+            PropertyIdentifier::EVENT_DETECTION_ENABLE,
             PropertyIdentifier::OUT_OF_SERVICE,
             PropertyIdentifier::UNITS,
             PropertyIdentifier::COV_INCREMENT,
@@ -228,7 +255,12 @@ impl BACnetObject for AnalogInputObject {
         Some(self.cov_increment)
     }
 
-    crate::impl_intrinsic_reporting!(event_detector, present_value, reliability);
+    crate::impl_intrinsic_reporting!(
+        event_detector,
+        present_value,
+        reliability,
+        event_detection_enable
+    );
 
     fn acknowledge_alarm(&mut self, transition_bit: u8) -> Result<(), bacnet_types::error::Error> {
         self.event_detector.acked_transitions |= transition_bit & 0x07;
@@ -246,5 +278,69 @@ impl BACnetObject for AnalogInputObject {
             || property == PropertyIdentifier::RELIABILITY
             || property == PropertyIdentifier::COV_INCREMENT
             || common::is_event_property_writable(property)
+            || property == PropertyIdentifier::EVENT_DETECTION_ENABLE
+    }
+}
+
+#[cfg(test)]
+mod detection_enable_reset_tests {
+    use super::*;
+
+    /// Regression guard for issue #123: once transitions populate timestamps and messages,
+    /// disabling event detection must still restore their Clause 13.2.2.1 initial conditions.
+    #[test]
+    fn ai_disabling_detection_resets_event_history() {
+        let mut ai = AnalogInputObject::new(1, "AI-1", 62).unwrap();
+        assert_eq!(
+            ai.read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
+                .unwrap(),
+            PropertyValue::Boolean(true)
+        );
+        ai.event_detector.event_state = bacnet_types::enums::EventState::HIGH_LIMIT;
+        ai.event_detector.acked_transitions = 0;
+        ai.event_detector.pending = Some(crate::event::PendingTransition {
+            state: bacnet_types::enums::EventState::HIGH_LIMIT,
+            remaining: 2,
+        });
+        ai.event_detector.fault_reliability = Some(1);
+        ai.event_time_stamps = [
+            BACnetTimeStamp::SequenceNumber(1),
+            BACnetTimeStamp::SequenceNumber(2),
+            BACnetTimeStamp::SequenceNumber(3),
+        ];
+        ai.event_message_texts = ["offnormal".into(), "fault".into(), "normal".into()];
+
+        ai.write_property(
+            PropertyIdentifier::EVENT_DETECTION_ENABLE,
+            None,
+            PropertyValue::Boolean(false),
+            None,
+        )
+        .unwrap();
+
+        assert_eq!(
+            ai.read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
+                .unwrap(),
+            PropertyValue::Boolean(false)
+        );
+        assert_eq!(
+            ai.event_detector.event_state,
+            bacnet_types::enums::EventState::NORMAL
+        );
+        assert_eq!(ai.event_detector.acked_transitions, 0b111);
+        assert!(ai.event_detector.pending.is_none());
+        assert!(ai.event_detector.fault_reliability.is_none());
+        assert_eq!(
+            ai.event_time_stamps,
+            [
+                BACnetTimeStamp::SequenceNumber(0),
+                BACnetTimeStamp::SequenceNumber(0),
+                BACnetTimeStamp::SequenceNumber(0),
+            ]
+        );
+        assert_eq!(
+            ai.event_message_texts,
+            [String::new(), String::new(), String::new()]
+        );
     }
 }
