+++ crates/bacnet-objects/src/analog/output.rs
@@ -24,6 +24,9 @@ pub struct AnalogOutputObject {
     /// Set to a positive value for delta-based filtering.
     cov_increment: f32,
     event_detector: OutOfRangeDetector,
+    /// Event_Detection_Enable (Clause 12.3). Clause 13.2.2.1: "If the
+    /// Event_Detection_Enable property is FALSE, then this state machine is not evaluated."
+    event_detection_enable: bool,
     reliability: u32,
     min_pres_value: Option<f32>,
     max_pres_value: Option<f32>,
@@ -49,6 +52,7 @@ impl AnalogOutputObject {
             relinquish_default: 0.0,
             cov_increment: 0.0,
             event_detector: OutOfRangeDetector::default(),
+            event_detection_enable: true,
             reliability: 0,
             min_pres_value: None,
             max_pres_value: None,
@@ -114,6 +118,9 @@ impl BACnetObject for AnalogOutputObject {
         }
         if let Some(result) = read_generic_event_properties!(self, property) {
             return result;
+        }
+        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
+            return Ok(PropertyValue::Boolean(self.event_detection_enable));
         }
         match property {
             p if p == PropertyIdentifier::OBJECT_TYPE => Ok(PropertyValue::Enumerated(
@@ -207,6 +214,25 @@ impl BACnetObject for AnalogOutputObject {
         {
             return result;
         }
+        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
+            if let PropertyValue::Boolean(v) = value {
+                self.event_detection_enable = v;
+                if !v {
+                    self.event_detector.event_state = bacnet_types::enums::EventState::NORMAL;
+                    self.event_detector.acked_transitions = 0b111;
+                    self.event_detector.pending = None;
+                    self.event_detector.fault_reliability = None;
+                    self.event_time_stamps = [
+                        BACnetTimeStamp::SequenceNumber(0),
+                        BACnetTimeStamp::SequenceNumber(0),
+                        BACnetTimeStamp::SequenceNumber(0),
+                    ];
+                    self.event_message_texts = [String::new(), String::new(), String::new()];
+                }
+                return Ok(());
+            }
+            return Err(common::invalid_data_type_error());
+        }
         if let Some(result) = write_analog_event_properties!(self, property, value) {
             return result;
         }
@@ -225,6 +251,7 @@ impl BACnetObject for AnalogOutputObject {
             PropertyIdentifier::PRESENT_VALUE,
             PropertyIdentifier::STATUS_FLAGS,
             PropertyIdentifier::EVENT_STATE,
+            PropertyIdentifier::EVENT_DETECTION_ENABLE,
             PropertyIdentifier::OUT_OF_SERVICE,
             PropertyIdentifier::UNITS,
             PropertyIdentifier::PRIORITY_ARRAY,
@@ -255,7 +282,12 @@ impl BACnetObject for AnalogOutputObject {
         Some(self.cov_increment)
     }
 
-    crate::impl_intrinsic_reporting!(event_detector, present_value, reliability);
+    crate::impl_intrinsic_reporting!(
+        event_detector,
+        present_value,
+        reliability,
+        event_detection_enable
+    );
 
     fn acknowledge_alarm(&mut self, transition_bit: u8) -> Result<(), bacnet_types::error::Error> {
         self.event_detector.acked_transitions |= transition_bit & 0x07;
@@ -274,5 +306,69 @@ impl BACnetObject for AnalogOutputObject {
             || property == PropertyIdentifier::RELIABILITY
             || property == PropertyIdentifier::COV_INCREMENT
             || common::is_event_property_writable(property)
+            || property == PropertyIdentifier::EVENT_DETECTION_ENABLE
+    }
+}
+
+#[cfg(test)]
+mod detection_enable_reset_tests {
+    use super::*;
+
+    /// Regression guard for issue #123: once transitions populate timestamps and messages,
+    /// disabling event detection must still restore their Clause 13.2.2.1 initial conditions.
+    #[test]
+    fn ao_disabling_detection_resets_event_history() {
+        let mut ao = AnalogOutputObject::new(1, "AO-1", 62).unwrap();
+        assert_eq!(
+            ao.read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
+                .unwrap(),
+            PropertyValue::Boolean(true)
+        );
+        ao.event_detector.event_state = bacnet_types::enums::EventState::HIGH_LIMIT;
+        ao.event_detector.acked_transitions = 0;
+        ao.event_detector.pending = Some(crate::event::PendingTransition {
+            state: bacnet_types::enums::EventState::HIGH_LIMIT,
+            remaining: 2,
+        });
+        ao.event_detector.fault_reliability = Some(1);
+        ao.event_time_stamps = [
+            BACnetTimeStamp::SequenceNumber(1),
+            BACnetTimeStamp::SequenceNumber(2),
+            BACnetTimeStamp::SequenceNumber(3),
+        ];
+        ao.event_message_texts = ["offnormal".into(), "fault".into(), "normal".into()];
+
+        ao.write_property(
+            PropertyIdentifier::EVENT_DETECTION_ENABLE,
+            None,
+            PropertyValue::Boolean(false),
+            None,
+        )
+        .unwrap();
+
+        assert_eq!(
+            ao.read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
+                .unwrap(),
+            PropertyValue::Boolean(false)
+        );
+        assert_eq!(
+            ao.event_detector.event_state,
+            bacnet_types::enums::EventState::NORMAL
+        );
+        assert_eq!(ao.event_detector.acked_transitions, 0b111);
+        assert!(ao.event_detector.pending.is_none());
+        assert!(ao.event_detector.fault_reliability.is_none());
+        assert_eq!(
+            ao.event_time_stamps,
+            [
+                BACnetTimeStamp::SequenceNumber(0),
+                BACnetTimeStamp::SequenceNumber(0),
+                BACnetTimeStamp::SequenceNumber(0),
+            ]
+        );
+        assert_eq!(
+            ao.event_message_texts,
+            [String::new(), String::new(), String::new()]
+        );
     }
 }
