+++ crates/bacnet-objects/src/binary/mod.rs
@@ -12,734 +12,9 @@ use crate::common::{
 use crate::event::{ChangeOfStateDetector, CommandFailureDetector};
 use crate::traits::BACnetObject;
 
-// ---------------------------------------------------------------------------
-// BinaryInput (type 3)
-// ---------------------------------------------------------------------------
-
-/// BACnet Binary Input object.
-///
-/// Read-only binary point. Present_Value is writable only when out-of-service.
-/// Uses Enumerated values: 0 = inactive, 1 = active.
-pub struct BinaryInputObject {
-    oid: ObjectIdentifier,
-    name: String,
-    description: String,
-    present_value: u32,
-    out_of_service: bool,
-    status_flags: StatusFlags,
-    /// Polarity: 0 = normal, 1 = reverse.
-    polarity: u32,
-    /// Reliability: 0 = NO_FAULT_DETECTED.
-    reliability: u32,
-    active_text: String,
-    inactive_text: String,
-    /// CHANGE_OF_STATE event detector.
-    event_detector: ChangeOfStateDetector,
-}
-
-impl BinaryInputObject {
-    pub fn new(instance: u32, name: impl Into<String>) -> Result<Self, Error> {
-        let oid = ObjectIdentifier::new(ObjectType::BINARY_INPUT, instance)?;
-        Ok(Self {
-            oid,
-            name: name.into(),
-            description: String::new(),
-            present_value: 0,
-            out_of_service: false,
-            status_flags: StatusFlags::empty(),
-            polarity: 0,
-            reliability: 0,
-            active_text: "Active".into(),
-            inactive_text: "Inactive".into(),
-            event_detector: ChangeOfStateDetector::default(),
-        })
-    }
-
-    /// Set the present value (used by application to update input state).
-    pub fn set_present_value(&mut self, value: u32) {
-        self.present_value = value;
-    }
-
-    /// Set the description string.
-    pub fn set_description(&mut self, desc: impl Into<String>) {
-        self.description = desc.into();
-    }
-}
-
-impl BACnetObject for BinaryInputObject {
-    fn object_identifier(&self) -> ObjectIdentifier {
-        self.oid
-    }
-
-    fn object_name(&self) -> &str {
-        &self.name
-    }
-
-    fn supports_cov(&self) -> bool {
-        true
-    }
-
-    crate::impl_intrinsic_reporting!(event_detector, present_value, reliability);
-
-    fn read_property(
-        &self,
-        property: PropertyIdentifier,
-        array_index: Option<u32>,
-    ) -> Result<PropertyValue, Error> {
-        if property == PropertyIdentifier::STATUS_FLAGS {
-            return Ok(common::compute_status_flags(
-                self.status_flags,
-                self.reliability,
-                self.out_of_service,
-                self.event_detector.event_state.to_raw(),
-            ));
-        }
-        if let Some(result) = read_common_properties!(self, property, array_index) {
-            return result;
-        }
-        match property {
-            p if p == PropertyIdentifier::OBJECT_TYPE => {
-                Ok(PropertyValue::Enumerated(ObjectType::BINARY_INPUT.to_raw()))
-            }
-            p if p == PropertyIdentifier::PRESENT_VALUE => {
-                Ok(PropertyValue::Enumerated(self.present_value))
-            }
-            p if p == PropertyIdentifier::EVENT_STATE => Ok(PropertyValue::Enumerated(
-                self.event_detector.event_state.to_raw(),
-            )),
-            p if p == PropertyIdentifier::POLARITY => Ok(PropertyValue::Enumerated(self.polarity)),
-            p if p == PropertyIdentifier::ACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.active_text.clone()))
-            }
-            p if p == PropertyIdentifier::INACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.inactive_text.clone()))
-            }
-            p if p == PropertyIdentifier::ALARM_VALUES => Ok(PropertyValue::List(
-                self.event_detector
-                    .alarm_values
-                    .iter()
-                    .map(|v| PropertyValue::Enumerated(*v))
-                    .collect(),
-            )),
-            p if p == PropertyIdentifier::EVENT_ENABLE => Ok(PropertyValue::BitString {
-                unused_bits: 5,
-                data: vec![self.event_detector.event_enable << 5],
-            }),
-            p if p == PropertyIdentifier::ACKED_TRANSITIONS => Ok(PropertyValue::BitString {
-                unused_bits: 5,
-                data: vec![self.event_detector.acked_transitions << 5],
-            }),
-            p if p == PropertyIdentifier::NOTIFICATION_CLASS => Ok(PropertyValue::Unsigned(
-                self.event_detector.notification_class as u64,
-            )),
-            p if p == PropertyIdentifier::EVENT_TIME_STAMPS => Ok(PropertyValue::List(vec![
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-            ])),
-            _ => Err(common::unknown_property_error()),
-        }
-    }
-
-    fn write_property(
-        &mut self,
-        property: PropertyIdentifier,
-        _array_index: Option<u32>,
-        value: PropertyValue,
-        _priority: Option<u8>,
-    ) -> Result<(), Error> {
-        if property == PropertyIdentifier::PRESENT_VALUE {
-            if !self.out_of_service {
-                return Err(common::write_access_denied_error());
-            }
-            if let PropertyValue::Enumerated(v) = value {
-                if v > 1 {
-                    return Err(common::value_out_of_range_error());
-                }
-                self.present_value = v;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::ACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.active_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::INACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.inactive_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if let Some(result) =
-            common::write_out_of_service(&mut self.out_of_service, property, &value)
-        {
-            return result;
-        }
-        if let Some(result) = common::write_object_name(&mut self.name, property, &value) {
-            return result;
-        }
-        if let Some(result) = common::write_description(&mut self.description, property, &value) {
-            return result;
-        }
-        Err(common::write_access_denied_error())
-    }
-
-    fn property_list(&self) -> Cow<'static, [PropertyIdentifier]> {
-        static PROPS: &[PropertyIdentifier] = &[
-            PropertyIdentifier::OBJECT_IDENTIFIER,
-            PropertyIdentifier::OBJECT_NAME,
-            PropertyIdentifier::DESCRIPTION,
-            PropertyIdentifier::OBJECT_TYPE,
-            PropertyIdentifier::PRESENT_VALUE,
-            PropertyIdentifier::STATUS_FLAGS,
-            PropertyIdentifier::EVENT_STATE,
-            PropertyIdentifier::OUT_OF_SERVICE,
-            PropertyIdentifier::POLARITY,
-            PropertyIdentifier::RELIABILITY,
-            PropertyIdentifier::ACTIVE_TEXT,
-            PropertyIdentifier::INACTIVE_TEXT,
-        ];
-        Cow::Borrowed(PROPS)
-    }
-
-    fn is_createable(&self) -> bool {
-        true
-    }
-
-    fn is_writable_property(&self, property: PropertyIdentifier) -> bool {
-        // Mirrors the BinaryInput `write_property` arms. EVENT_ENABLE,
-        // NOTIFICATION_CLASS, and ACKED_TRANSITIONS are read-only on BI
-        // (no write arms), so they are intentionally excluded.
-        common::is_common_writable(property)
-            || property == PropertyIdentifier::PRESENT_VALUE
-            || property == PropertyIdentifier::ACTIVE_TEXT
-            || property == PropertyIdentifier::INACTIVE_TEXT
-    }
-}
-
-// ---------------------------------------------------------------------------
-// BinaryOutput (type 4)
-// ---------------------------------------------------------------------------
-
-/// BACnet Binary Output object.
-///
-/// Commandable binary output with 16-level priority array.
-/// Uses Enumerated values: 0 = inactive, 1 = active.
-pub struct BinaryOutputObject {
-    oid: ObjectIdentifier,
-    name: String,
-    description: String,
-    present_value: u32,
-    feedback_value: u32,
-    out_of_service: bool,
-    status_flags: StatusFlags,
-    priority_array: [Option<u32>; 16],
-    relinquish_default: u32,
-    /// Polarity: 0 = normal, 1 = reverse.
-    polarity: u32,
-    /// Reliability: 0 = NO_FAULT_DETECTED.
-    reliability: u32,
-    event_detection_enable: bool,
-    active_text: String,
-    inactive_text: String,
-    /// COMMAND_FAILURE event detector.
-    event_detector: CommandFailureDetector,
-    /// Value source tracking (optional per spec — exposed via VALUE_SOURCE property).
-    value_source: common::ValueSourceTracking,
-}
-
-impl BinaryOutputObject {
-    pub fn new(instance: u32, name: impl Into<String>) -> Result<Self, Error> {
-        let oid = ObjectIdentifier::new(ObjectType::BINARY_OUTPUT, instance)?;
-        Ok(Self {
-            oid,
-            name: name.into(),
-            description: String::new(),
-            present_value: 0,
-            feedback_value: 0,
-            out_of_service: false,
-            status_flags: StatusFlags::empty(),
-            priority_array: [None; 16],
-            relinquish_default: 0,
-            polarity: 0,
-            reliability: 0,
-            event_detection_enable: false,
-            active_text: "Active".into(),
-            inactive_text: "Inactive".into(),
-            event_detector: CommandFailureDetector::default(),
-            value_source: common::ValueSourceTracking::default(),
-        })
-    }
-
-    /// Set the description string.
-    pub fn set_description(&mut self, desc: impl Into<String>) {
-        self.description = desc.into();
-    }
-
-    fn recalculate_present_value(&mut self) {
-        self.present_value =
-            common::recalculate_from_priority_array(&self.priority_array, self.relinquish_default);
-    }
-}
-
-impl BACnetObject for BinaryOutputObject {
-    fn object_identifier(&self) -> ObjectIdentifier {
-        self.oid
-    }
-
-    fn object_name(&self) -> &str {
-        &self.name
-    }
-
-    fn supports_cov(&self) -> bool {
-        true
-    }
-
-    crate::impl_intrinsic_reporting!(
-        event_detector,
-        present_value,
-        feedback_value,
-        reliability,
-        event_detection_enable
-    );
-
-    fn read_property(
-        &self,
-        property: PropertyIdentifier,
-        array_index: Option<u32>,
-    ) -> Result<PropertyValue, Error> {
-        if property == PropertyIdentifier::STATUS_FLAGS {
-            return Ok(common::compute_status_flags(
-                self.status_flags,
-                self.reliability,
-                self.out_of_service,
-                self.event_detector.event_state.to_raw(),
-            ));
-        }
-        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
-            return Ok(PropertyValue::Boolean(self.event_detection_enable));
-        }
-        if let Some(result) = read_common_properties!(self, property, array_index) {
-            return result;
-        }
-        if let Some(result) = read_generic_event_properties!(self, property) {
-            return result;
-        }
-        match property {
-            p if p == PropertyIdentifier::OBJECT_TYPE => Ok(PropertyValue::Enumerated(
-                ObjectType::BINARY_OUTPUT.to_raw(),
-            )),
-            p if p == PropertyIdentifier::PRESENT_VALUE => {
-                Ok(PropertyValue::Enumerated(self.present_value))
-            }
-            p if p == PropertyIdentifier::FEEDBACK_VALUE => {
-                Ok(PropertyValue::Enumerated(self.feedback_value))
-            }
-            p if p == PropertyIdentifier::PRIORITY_ARRAY => {
-                common::read_priority_array!(self, array_index, PropertyValue::Enumerated)
-            }
-            p if p == PropertyIdentifier::RELINQUISH_DEFAULT => {
-                Ok(PropertyValue::Enumerated(self.relinquish_default))
-            }
-            p if p == PropertyIdentifier::CURRENT_COMMAND_PRIORITY => {
-                Ok(common::current_command_priority(&self.priority_array))
-            }
-            p if p == PropertyIdentifier::VALUE_SOURCE => {
-                Ok(self.value_source.value_source.clone())
-            }
-            p if p == PropertyIdentifier::LAST_COMMAND_TIME => Ok(PropertyValue::Unsigned(
-                match self.value_source.last_command_time {
-                    BACnetTimeStamp::SequenceNumber(n) => n,
-                    _ => 0,
-                },
-            )),
-            p if p == PropertyIdentifier::POLARITY => Ok(PropertyValue::Enumerated(self.polarity)),
-            p if p == PropertyIdentifier::ACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.active_text.clone()))
-            }
-            p if p == PropertyIdentifier::INACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.inactive_text.clone()))
-            }
-            p if p == PropertyIdentifier::EVENT_TIME_STAMPS => Ok(PropertyValue::List(vec![
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-            ])),
-            _ => Err(common::unknown_property_error()),
-        }
-    }
-
-    fn write_property(
-        &mut self,
-        property: PropertyIdentifier,
-        array_index: Option<u32>,
-        value: PropertyValue,
-        priority: Option<u8>,
-    ) -> Result<(), Error> {
-        common::write_priority_array_direct!(self, property, array_index, value, |v| {
-            if let PropertyValue::Enumerated(e) = v {
-                if e > 1 {
-                    Err(common::value_out_of_range_error())
-                } else {
-                    Ok(e)
-                }
-            } else {
-                Err(common::invalid_data_type_error())
-            }
-        });
-        if property == PropertyIdentifier::PRESENT_VALUE {
-            return common::write_priority_array!(self, value, priority, |v| {
-                if let PropertyValue::Enumerated(e) = v {
-                    if e > 1 {
-                        Err(common::value_out_of_range_error())
-                    } else {
-                        Ok(e)
-                    }
-                } else {
-                    Err(common::invalid_data_type_error())
-                }
-            });
-        }
-        if property == PropertyIdentifier::FEEDBACK_VALUE {
-            if let PropertyValue::Enumerated(e) = value {
-                if e > 1 {
-                    return Err(common::value_out_of_range_error());
-                }
-                self.feedback_value = e;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::ACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.active_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::INACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.inactive_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::EVENT_DETECTION_ENABLE {
-            if let PropertyValue::Boolean(v) = value {
-                self.event_detection_enable = v;
-                if !v {
-                    self.event_detector.event_state = bacnet_types::enums::EventState::NORMAL;
-                    self.event_detector.acked_transitions = 0b111;
-                    self.event_detector.pending = None;
-                    self.event_detector.fault_reliability = None;
-                }
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if let Some(result) = write_generic_event_properties!(self, property, value) {
-            return result;
-        }
-        if let Some(result) =
-            common::write_out_of_service(&mut self.out_of_service, property, &value)
-        {
-            return result;
-        }
-        if let Some(result) = common::write_object_name(&mut self.name, property, &value) {
-            return result;
-        }
-        if let Some(result) = common::write_description(&mut self.description, property, &value) {
-            return result;
-        }
-        Err(common::write_access_denied_error())
-    }
-
-    fn property_list(&self) -> Cow<'static, [PropertyIdentifier]> {
-        static PROPS: &[PropertyIdentifier] = &[
-            PropertyIdentifier::OBJECT_IDENTIFIER,
-            PropertyIdentifier::OBJECT_NAME,
-            PropertyIdentifier::DESCRIPTION,
-            PropertyIdentifier::OBJECT_TYPE,
-            PropertyIdentifier::PRESENT_VALUE,
-            PropertyIdentifier::FEEDBACK_VALUE,
-            PropertyIdentifier::STATUS_FLAGS,
-            PropertyIdentifier::EVENT_STATE,
-            PropertyIdentifier::EVENT_DETECTION_ENABLE,
-            PropertyIdentifier::EVENT_ENABLE,
-            PropertyIdentifier::TIME_DELAY,
-            PropertyIdentifier::NOTIFY_TYPE,
-            PropertyIdentifier::NOTIFICATION_CLASS,
-            PropertyIdentifier::ACKED_TRANSITIONS,
-            PropertyIdentifier::OUT_OF_SERVICE,
-            PropertyIdentifier::PRIORITY_ARRAY,
-            PropertyIdentifier::RELINQUISH_DEFAULT,
-            PropertyIdentifier::CURRENT_COMMAND_PRIORITY,
-            PropertyIdentifier::POLARITY,
-            PropertyIdentifier::RELIABILITY,
-            PropertyIdentifier::ACTIVE_TEXT,
-            PropertyIdentifier::INACTIVE_TEXT,
-        ];
-        Cow::Borrowed(PROPS)
-    }
-
-    fn is_createable(&self) -> bool {
-        true
-    }
-
-    fn is_writable_property(&self, property: PropertyIdentifier) -> bool {
-        // Mirrors the BinaryOutput `write_property` arms: commandable
-        // (PRIORITY_ARRAY + PRESENT_VALUE) + common + text properties.
-        common::is_commandable_property_writable(property)
-            || common::is_common_writable(property)
-            || common::is_generic_event_property_writable(property)
-            || property == PropertyIdentifier::FEEDBACK_VALUE
-            || property == PropertyIdentifier::ACTIVE_TEXT
-            || property == PropertyIdentifier::INACTIVE_TEXT
-            || property == PropertyIdentifier::EVENT_DETECTION_ENABLE
-    }
-}
-
-// ---------------------------------------------------------------------------
-// BinaryValue (type 5)
-// ---------------------------------------------------------------------------
-
-/// BACnet Binary Value object.
-///
-/// Commandable binary value with 16-level priority array.
-/// Uses Enumerated values: 0 = inactive, 1 = active.
-pub struct BinaryValueObject {
-    oid: ObjectIdentifier,
-    name: String,
-    description: String,
-    present_value: u32, // 0 = inactive, 1 = active
-    out_of_service: bool,
-    status_flags: StatusFlags,
-    priority_array: [Option<u32>; 16],
-    relinquish_default: u32,
-    /// Reliability: 0 = NO_FAULT_DETECTED.
-    reliability: u32,
-    active_text: String,
-    inactive_text: String,
-    /// CHANGE_OF_STATE event detector.
-    event_detector: ChangeOfStateDetector,
-    /// Value source tracking (optional per spec — exposed via VALUE_SOURCE property).
-    value_source: common::ValueSourceTracking,
-}
-
-impl BinaryValueObject {
-    /// Create a new Binary Value object.
-    pub fn new(instance: u32, name: impl Into<String>) -> Result<Self, Error> {
-        let oid = ObjectIdentifier::new(ObjectType::BINARY_VALUE, instance)?;
-        Ok(Self {
-            oid,
-            name: name.into(),
-            description: String::new(),
-            present_value: 0, // inactive
-            out_of_service: false,
-            status_flags: StatusFlags::empty(),
-            priority_array: [None; 16],
-            relinquish_default: 0,
-            reliability: 0,
-            active_text: "Active".into(),
-            inactive_text: "Inactive".into(),
-            event_detector: ChangeOfStateDetector::default(),
-            value_source: common::ValueSourceTracking::default(),
-        })
-    }
-
-    /// Set the description string.
-    pub fn set_description(&mut self, desc: impl Into<String>) {
-        self.description = desc.into();
-    }
-
-    fn recalculate_present_value(&mut self) {
-        self.present_value =
-            common::recalculate_from_priority_array(&self.priority_array, self.relinquish_default);
-    }
-}
-
-impl BACnetObject for BinaryValueObject {
-    fn object_identifier(&self) -> ObjectIdentifier {
-        self.oid
-    }
-
-    fn object_name(&self) -> &str {
-        &self.name
-    }
-
-    fn supports_cov(&self) -> bool {
-        true
-    }
-
-    crate::impl_intrinsic_reporting!(event_detector, present_value, reliability);
-
-    fn read_property(
-        &self,
-        property: PropertyIdentifier,
-        array_index: Option<u32>,
-    ) -> Result<PropertyValue, Error> {
-        if property == PropertyIdentifier::STATUS_FLAGS {
-            return Ok(common::compute_status_flags(
-                self.status_flags,
-                self.reliability,
-                self.out_of_service,
-                self.event_detector.event_state.to_raw(),
-            ));
-        }
-        if let Some(result) = read_common_properties!(self, property, array_index) {
-            return result;
-        }
-        match property {
-            p if p == PropertyIdentifier::OBJECT_TYPE => {
-                Ok(PropertyValue::Enumerated(ObjectType::BINARY_VALUE.to_raw()))
-            }
-            p if p == PropertyIdentifier::PRESENT_VALUE => {
-                Ok(PropertyValue::Enumerated(self.present_value))
-            }
-            p if p == PropertyIdentifier::EVENT_STATE => Ok(PropertyValue::Enumerated(
-                self.event_detector.event_state.to_raw(),
-            )),
-            p if p == PropertyIdentifier::PRIORITY_ARRAY => {
-                common::read_priority_array!(self, array_index, PropertyValue::Enumerated)
-            }
-            p if p == PropertyIdentifier::RELINQUISH_DEFAULT => {
-                Ok(PropertyValue::Enumerated(self.relinquish_default))
-            }
-            p if p == PropertyIdentifier::CURRENT_COMMAND_PRIORITY => {
-                Ok(common::current_command_priority(&self.priority_array))
-            }
-            p if p == PropertyIdentifier::VALUE_SOURCE => {
-                Ok(self.value_source.value_source.clone())
-            }
-            p if p == PropertyIdentifier::LAST_COMMAND_TIME => Ok(PropertyValue::Unsigned(
-                match self.value_source.last_command_time {
-                    BACnetTimeStamp::SequenceNumber(n) => n,
-                    _ => 0,
-                },
-            )),
-            p if p == PropertyIdentifier::ACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.active_text.clone()))
-            }
-            p if p == PropertyIdentifier::INACTIVE_TEXT => {
-                Ok(PropertyValue::CharacterString(self.inactive_text.clone()))
-            }
-            p if p == PropertyIdentifier::EVENT_ENABLE => Ok(PropertyValue::BitString {
-                unused_bits: 5,
-                data: vec![self.event_detector.event_enable << 5],
-            }),
-            p if p == PropertyIdentifier::ACKED_TRANSITIONS => Ok(PropertyValue::BitString {
-                unused_bits: 5,
-                data: vec![self.event_detector.acked_transitions << 5],
-            }),
-            p if p == PropertyIdentifier::NOTIFICATION_CLASS => Ok(PropertyValue::Unsigned(
-                self.event_detector.notification_class as u64,
-            )),
-            p if p == PropertyIdentifier::EVENT_TIME_STAMPS => Ok(PropertyValue::List(vec![
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-                PropertyValue::Unsigned(0),
-            ])),
-            _ => Err(common::unknown_property_error()),
-        }
-    }
-
-    fn write_property(
-        &mut self,
-        property: PropertyIdentifier,
-        array_index: Option<u32>,
-        value: PropertyValue,
-        priority: Option<u8>,
-    ) -> Result<(), Error> {
-        common::write_priority_array_direct!(self, property, array_index, value, |v| {
-            if let PropertyValue::Enumerated(e) = v {
-                if e > 1 {
-                    Err(common::value_out_of_range_error())
-                } else {
-                    Ok(e)
-                }
-            } else {
-                Err(common::invalid_data_type_error())
-            }
-        });
-        if property == PropertyIdentifier::PRESENT_VALUE {
-            return common::write_priority_array!(self, value, priority, |v| {
-                if let PropertyValue::Enumerated(e) = v {
-                    if e > 1 {
-                        Err(common::value_out_of_range_error())
-                    } else {
-                        Ok(e)
-                    }
-                } else {
-                    Err(common::invalid_data_type_error())
-                }
-            });
-        }
-        if property == PropertyIdentifier::ACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.active_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if property == PropertyIdentifier::INACTIVE_TEXT {
-            if let PropertyValue::CharacterString(s) = value {
-                self.inactive_text = s;
-                return Ok(());
-            }
-            return Err(common::invalid_data_type_error());
-        }
-        if let Some(result) =
-            common::write_out_of_service(&mut self.out_of_service, property, &value)
-        {
-            return result;
-        }
-        if let Some(result) = common::write_object_name(&mut self.name, property, &value) {
-            return result;
-        }
-        if let Some(result) = common::write_description(&mut self.description, property, &value) {
-            return result;
-        }
-        Err(common::write_access_denied_error())
-    }
-
-    fn property_list(&self) -> Cow<'static, [PropertyIdentifier]> {
-        static PROPS: &[PropertyIdentifier] = &[
-            PropertyIdentifier::OBJECT_IDENTIFIER,
-            PropertyIdentifier::OBJECT_NAME,
-            PropertyIdentifier::DESCRIPTION,
-            PropertyIdentifier::OBJECT_TYPE,
-            PropertyIdentifier::PRESENT_VALUE,
-            PropertyIdentifier::STATUS_FLAGS,
-            PropertyIdentifier::EVENT_STATE,
-            PropertyIdentifier::OUT_OF_SERVICE,
-            PropertyIdentifier::PRIORITY_ARRAY,
-            PropertyIdentifier::RELINQUISH_DEFAULT,
-            PropertyIdentifier::CURRENT_COMMAND_PRIORITY,
-            PropertyIdentifier::RELIABILITY,
-            PropertyIdentifier::ACTIVE_TEXT,
-            PropertyIdentifier::INACTIVE_TEXT,
-        ];
-        Cow::Borrowed(PROPS)
-    }
-
-    fn is_createable(&self) -> bool {
-        true
-    }
-
-    fn is_writable_property(&self, property: PropertyIdentifier) -> bool {
-        // Mirrors the BinaryValue `write_property` arms. Same set as
-        // BinaryOutput (commandable + common + text properties).
-        common::is_commandable_property_writable(property)
-            || common::is_common_writable(property)
-            || property == PropertyIdentifier::ACTIVE_TEXT
-            || property == PropertyIdentifier::INACTIVE_TEXT
-    }
-}
-
-#[cfg(test)]
-mod tests;
+mod input;
+mod output;
+mod value;
+pub use input::*;
+pub use output::*;
+pub use value::*;
