+++ crates/bacnet-objects/src/event.rs
@@ -907,6 +907,9 @@ impl CommandFailureDetector {
 /// Reliability written by any route — the server's fault detector, a local
 /// write, or a network write — reaches event-state-detection without each route
 /// needing to notify it.
+///
+/// The gated arms enforce ASHRAE 135-2020 Clause 13.2.2.1: "If the
+/// Event_Detection_Enable property is FALSE, then this state machine is not evaluated."
 #[macro_export]
 macro_rules! impl_intrinsic_reporting {
     (
@@ -938,24 +941,31 @@ macro_rules! impl_intrinsic_reporting {
             )
         }
     };
-    // There is deliberately no four-field arm — no feedback-driven detector without an
-    // Event_Detection_Enable gate. One existed briefly during this change and became unused
-    // once Binary Output and Multi-state Output moved to the gated form. Leaving it exported
-    // would have offered downstream implementors a supported way to wire a three-input
-    // detector with detection permanently on, which is the exact defect this gate was added
-    // to fix: Clauses 12.7 and 12.19 make intrinsic reporting optional ("required if the
-    // object supports intrinsic reporting"), and an ungated detector makes it mandatory.
-    ($detector_field:ident, $present_value_field:ident, $reliability_field:ident) => {
+    // Gated two-input detector delegation for intrinsic-reporting object types without a
+    // feedback value.
+    (
+        $detector_field:ident,
+        $present_value_field:ident,
+        $reliability_field:ident,
+        $event_detection_enable_field:ident
+    ) => {
         fn evaluate_intrinsic_reporting(&mut self) -> Option<$crate::event::TransitionOutcome> {
+            if !self.$event_detection_enable_field {
+                return None;
+            }
             self.$detector_field
                 .probe(self.$present_value_field, self.$reliability_field)
         }
 
         fn tick_intrinsic_reporting(&mut self) -> Option<$crate::event::TransitionOutcome> {
+            if !self.$event_detection_enable_field {
+                return None;
+            }
             self.$detector_field
                 .tick(self.$present_value_field, self.$reliability_field)
         }
-    };
+    }; // There is deliberately no ungated arm. Exporting one would let downstream
+       // implementors wire event-state detection permanently on despite Clause 13.2.2.1.
 }
 
 #[cfg(test)]
